Move `Index` into it's own file
